properly imports body field
previously the body field could not be positioned by display suite
for some reason. Attempting to import the field using the existing
method caused an error since it was attempting to create a field
which already existed. However features module had a snippet which
checks if a field exists, and only creates an instance of it in
that case.